+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 - A Docker Compose example using the community Nextcloud image is now available in the repo.
 - **Feature**: Manual upload to folder is now available in the Android app.
+- **Feature**: Allow sorting albums in ascending and descending direction.
 
 ## [v7.0.2] - 2024-03-19